trying to set up socket connection
