+++ .gitignore
@@ -80,3 +80,4 @@ CHANGELOG.md
 node_modules
 # Optional Idea IDE
 .idea
+/pokemap.bundle.js
