+++ module/main.js
@@ -3,6 +3,7 @@
 var L = require('leaflet');
 require('leaflet/dist/leaflet.css');
 require('../style.css');
+var DataService = require('./DataService.js');
 
 // options - {
 //     coordinates: {       // optional
@@ -12,6 +13,7 @@ require('../style.css');
 //     zoomLevel: 10,       // optional
 //     timeRange: 1,        // optional
 //     apiEndpoint: 'URI'   // mandatory
+//     webSocketEndPoint: 'URI' //mandatory
 // }
 
 (function () {
@@ -23,6 +25,7 @@ require('../style.css');
         var zoomLevel = options.zoomLevel;
         var timeRange = options.timeRange;
         var apiEndpoint = options.apiEndpoint;
+        var socketEndPoint = options.webSocketEndPoint;
         var tileLayer = options.tileLayer;
         var tileLayerOptions;
 
@@ -36,6 +39,10 @@ require('../style.css');
 
         if (!apiEndpoint) {
             throw new Error('Fatal: apiEndpoint not defined');
+        }
+
+        if (!socketEndPoint) {
+            throw new Error('Fatal: socketEndPoint not defined');
         }
 
         if (!tileLayer) {
@@ -62,7 +69,7 @@ require('../style.css');
         var eventHandlers = {};
         var mymap = null;
         var pokemonLayer = null;
-        var dataService = new DataService(apiEndpoint);
+        var dataService = new DataService(apiEndpoint, socketEndPoint);
 
         initMap();
 
@@ -232,169 +239,6 @@ require('../style.css');
             return marker;
 
         }
-
-    }
-
-    function DataService(apiEndpoint) {
-
-        var self = this;
-
-        var dbService = {
-
-            getPastData: function (location, callback) {
-
-                var locationFrom = location.from.lng + ',' + location.from.lat;
-                var locationTo = location.to.lng + ',' + location.to.lat;
-
-                var xhr = new XMLHttpRequest();
-                var url = apiEndpoint + '/api/pokemon/sighting/coordinates/from/' + locationFrom + '/to/' + locationTo;
-                xhr.open("GET", url, true);
-                xhr.onreadystatechange = function () {
-
-                    if (xhr.readyState === 4 && xhr.status === 200) {
-
-                        var json = JSON.parse(xhr.responseText);
-
-                        callback(json);
-
-                    } else {
-
-                    }
-                };
-
-                xhr.send();
-
-            },
-
-            //supposing that we could get the predicted data through the same api
-            getPredictedData: function (location, callback) {
-
-                var locationFrom = location.from.lng + ',' + location.from.lat;
-                var locationTo = location.to.lng + ',' + location.to.lat;
-
-                var xhr = new XMLHttpRequest();
-                var url = apiEndpoint + 'api/pokemon/sighting/coordinates/from/' + locationFrom + '/to/' + locationTo;
-                xhr.open("GET", url, true);
-                xhr.onreadystatechange = function () {
-
-                    if (xhr.readyState === 4 && xhr.status === 200) {
-
-                        var json = JSON.parse(xhr.responseText);
-
-                        callback(json);
-
-                    } else {
-
-                    }
-                };
-
-                xhr.send();
-
-            },
-
-            getPokemonDetailsById: function (id, callback) {
-
-                var xhr = new XMLHttpRequest();
-                var url = apiEndpoint + '/api/pokemon/id/' + id;
-                xhr.open("GET", url, true);
-                xhr.onreadystatechange = function () {
-
-                    if (xhr.readyState === 4 && xhr.status === 200) {
-
-                        var json = JSON.parse(xhr.responseText);
-
-                        callback(json);
-
-                    } else {
-
-                    }
-                };
-
-                xhr.send();
-
-            },
-
-            getPokemonDataByTimeRange: function (from, to, callback) {
-
-                //The way the URL is requested is a bit different from what Catch em All group was thinking. Maybe we
-                //need to talk to the Data team to change this API is requested (just in seconds or minutes before and after.
-
-                //TODO: To be rechecked. the range is not clear how should it be specified. Currently not working.
-
-                var currentTime = new Date();
-                var startTimeStamp = new Date(currentTime.getTime() + 1000 * from);
-                var startTimeStampString = startTimeStamp.toUTCString();
-                var range = to - from + 's';
-
-                var xhr = new XMLHttpRequest();
-                var url = apiEndpoint + 'api/pokemon/sighting/ts/' + startTimeStampString + '/range/' + range;
-                xhr.open("GET", url, true);
-                xhr.onreadystatechange = function () {
-
-                    if (xhr.readyState === 4 && xhr.status === 200) {
-
-                        var json = JSON.parse(xhr.responseText);
-
-                        callback(json);
-
-                    } else {
-
-                    }
-                };
-
-                xhr.send();
-
-            }
-        };
-
-        function twitterService() {
-            function getTwitterData() {
-
-            }
-        }
-
-        self.getApiEndpointURL = function () {
-            return apiEndpoint;
-        };
-
-        self.getData = function (bounds, updateCallback) {
-
-            dbService.getPastData(bounds, updateCallback);
-            return;
-
-
-            if (timeRange.start < 0 && timeRange.end < 0) {
-
-                //get past data from database
-                var pokemons = dbService.getPastData(bounds, updateCallback);
-                return pokemons;
-
-            } else {
-
-                if (timeRange.start > 0 && timeRange.end > 0) {
-
-                    //get predictions from database
-                    var pokemons = dbService.getPredictedData(from, to, updateCallback);
-                    return pokemons;
-
-                } else {
-
-                    //get data from database
-                    //get data from twitter via sockets
-                    var pokemons = dbService.getPastData(bounds, updateCallback);
-
-                    pokemons.push(dbService.getPredictedData(bounds, updateCallback));
-                    return pokemons;
-
-                }
-            }
-
-        };
-
-        self.getPokemonDetailsById = function (id, callback) {
-
-            dbService.getPokemonDetailsById(id, callback);
-        };
 
     }
