+++ package.json
@@ -23,6 +23,7 @@
     "browserify": "^13.1.0",
     "browserify-css": "^0.9.2",
     "es6ify": "1.6.0",
-    "leaflet": "^0.7.7"
+    "leaflet": "^0.7.7",
+    "socket.io-client": "^1.4.8"
   }
 }
