+++ script.js
@@ -1,7 +1,8 @@
 var htmlElement = 'mapid';
 var options = {
 
-    apiEndpoint: 'http://pokedata.c4e3f8c7.svc.dockerapp.io:65014'
+    apiEndpoint: 'http://pokedata.c4e3f8c7.svc.dockerapp.io:65014',
+    webSocketEndPoint: 'http://pokedata.c4e3f8c7.svc.dockerapp.io:65024/sentiment'
 
 };
